requirements.txt
torch>=1.8.0
gym>=0.21.0
numpy>=1.19.2
pyyaml>=5.4
requests>=2.25.1